Use nested suites with single docker container stack.
