+++ src/test/scala/com/exasol/cloudetl/parquet/BaseParquetReaderTest.scala
@@ -11,15 +11,16 @@ import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{Path => HPath}
 import org.apache.hadoop.fs.FileSystem
 import org.apache.parquet.example.data.Group
-import org.apache.parquet.example.data.GroupWriter
 import org.apache.parquet.hadoop.ParquetWriter
-import org.apache.parquet.hadoop.api.WriteSupport
-import org.apache.parquet.io.api.RecordConsumer
-import org.apache.parquet.schema._
+import org.apache.parquet.schema.MessageType
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.funsuite.AnyFunSuite
 
-trait BaseParquetReaderTest extends AnyFunSuite with BeforeAndAfterEach with TestFileManager {
+trait BaseParquetReaderTest
+    extends AnyFunSuite
+    with BeforeAndAfterEach
+    with TestFileManager
+    with ParquetTestDataWriter {
 
   private[this] var conf: Configuration = _
   private[this] var fileSystem: FileSystem = _
@@ -45,37 +46,11 @@ trait BaseParquetReaderTest extends AnyFunSuite with BeforeAndAfterEach with Tes
   }
 
   protected final def getRecords(): Seq[Row] =
-    ParquetSource(path, conf, fileSystem)
-      .stream()
-      .toSeq
+    ParquetSource(path, conf, fileSystem).stream().toSeq
 
   protected final def getParquetWriter(
     schema: MessageType,
-    dictionaryEncoding: Boolean
+    encoding: Boolean
   ): ParquetWriter[Group] =
-    BaseGroupWriterBuilder(path, schema)
-      .withDictionaryEncoding(dictionaryEncoding)
-      .build()
-
-  private[this] case class BaseGroupWriteSupport(schema: MessageType)
-      extends WriteSupport[Group] {
-    var writer: GroupWriter = null
-
-    override def prepareForWrite(recordConsumer: RecordConsumer): Unit =
-      writer = new GroupWriter(recordConsumer, schema)
-
-    override def init(configuration: Configuration): WriteSupport.WriteContext =
-      new WriteSupport.WriteContext(schema, new java.util.HashMap[String, String]())
-
-    override def write(record: Group): Unit =
-      writer.write(record)
-  }
-
-  private[this] case class BaseGroupWriterBuilder(path: HPath, schema: MessageType)
-      extends ParquetWriter.Builder[Group, BaseGroupWriterBuilder](path) {
-    override def getWriteSupport(conf: Configuration): WriteSupport[Group] =
-      BaseGroupWriteSupport(schema)
-    override def self(): BaseGroupWriterBuilder = this
-  }
-
+    getParquetWriter(path, schema, encoding)
 }
